requirements.txt
gradio==4.44.1
gradio-client==1.3.0
fastapi==0.114.1
pandas==2.2.2
openpyxl==3.1.5
gspread==6.0.2
google-auth==2.29.0
python-dotenv==1.0.1
pydantic==2.4.2